Refactor: Split test_window_partial_constant_and_set_monotonicity from enforce_sorting
